### 2023-11-10 #6051 @editorialbot
[PRE REVIEW]: seesus: a social, environmental, and economic sustainability classifier for Python
**Submitting author:** @caimeng2 (<a href="http://orcid.org/0000-0002-8318-572X">Meng Cai</a>)
**Repository:** https://github.com/caimeng2/seesus
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.0
