+++ paper/paper.bib
@@ -115,6 +115,7 @@
 @article{UN_2015,
  title={Transforming our world: The 2030 agenda for sustainable development},
  journal={Resolution adopted by the General Assembly},
+ DOI={10.5040/9781509934058.0025},
  author={UN},
  year={2015}
 }
