+++ .github/workflows/draft-pdf.yml
@@ -21,8 +21,8 @@ jobs:
           # PDF. Note, this should be the same directory as the input
           # paper.md
           path: paper/paper.pdf
-      ## suspend this process with GitHub commits by commenting the following code
-      - name: save pdf to repo
-        uses: stefanzweifel/git-auto-commit-action@v4
-        with:
-          commit_message: Saved new PDF of paper
+      # ## suspend this process with GitHub commits by commenting the following code
+      # - name: save pdf to repo
+      #   uses: stefanzweifel/git-auto-commit-action@v4
+      #   with:
+      #     commit_message: Saved new PDF of paper
