### 2024-01-03 fe9920c
add print(result)
