### 2024-01-07 #6051 @luyuhao0326
@oliviaguest thanks for the invitation. I can certainly review this submission. However, I am not proficient in Python to provide constructive feedback on coding. Please let me know if JOSS would like me to review other aspects of the submission. Happy 2024.
