### 2024-01-22 #6244 @luyuhao0326
Thanks for the invitation and below are my reviews on installation, and software paper (most of my comments will be related to the paper since I am not a proficient Python user).

**Installation**
- It would be great to include more detailed installation instructions for users who are not so familiar with Python (e.g., me and many others who will be potentially benefiting from this work) and/or GitHub.
- That being said, I am unable to install and run this package on my machine. I will be happy to do so if one of the authors can help me install the package.

**Software paper**

- The idea is novel and I can see this work being useful in many domains. I have one question regarding the example use cases listed in the paper: the paper claims that seesus can be used in "_label academic publications_" and "_large-scale scans of planning documents_". However, the example in README.md only shows how seesus _evaluates individual sentences_ which can cause potential misinterpretation and biased results as the context of "academic publications" and "planning documents" will likely be missing when being evaluated sentence by sentence. Example 3 provided [here](https://github.com/caimeng2/seesus/blob/main/example.ipynb), for example is not really a paragraph.

- The statement of need is clear but a bit thin. Although I can appreciate JOSS is a more software-focused journal, it would still be great to provide some context on the current status of text mining/classification on UN-SDG and why it is important to for example, "_quantify which dimension of sustainability receives the most attention_"

- Accuracy of 75.5% is decent but is not particularly high. Considering the evaluation method is from a different package, it would be great if the authors can provide a statement (or even better specific development plans) on how to improve the accuracy and/or usabililty of future text-mining on SDG.
