+++ README.md
@@ -15,6 +15,8 @@
 
 # Installation
 
+You can install `seesus` from PyPI by inputting the following command in your terminal.
+
 `pip install seesus`
 
 
