### 2024-01-23 c77e33e
add example of an abstract
