### 2024-01-23 c8162fe
minor wording edits
