+++ README.md
@@ -23,7 +23,7 @@ You can install `seesus` from PyPI by inputting the following command in your te
 ## Example
 
 
-### Example 1: analyzing an individual sentence
+### Analyzing an individual sentence
 
 ```python
 from seesus import SeeSus
@@ -51,7 +51,7 @@ print(result1.target_desc)
 print(result1.see)
 ```
 
-### Example 2: analyzing a paragraph 
+### Analyzing a paragraph /a longer document
 
 To achieve the best results, it is recommended to split a paragraph or a whole document into individual sentences (i.e., using individual sentences as the basic unit for `seesus` to analyze). This can be done by tools such as `nltk.tokenize` and `re.split`.
 
