### 2024-01-24 #6244 @caimeng2
Hi @luyuhao0326,

Thank you very much for taking the time to review `seesus`. We appreciate your helpful feedback. Please find our point-by-point responses below.

**Installation**

> It would be great to include more detailed installation instructions for users who are not so familiar with Python (e.g., me and many others who will be potentially benefiting from this work) and/or GitHub.

Thank you for your suggestion. `seesus` is indeed a Python-based software that requires basic knowledge of Python programming. To simplify the installation process, we chose to publish `seesus` to PyPI and use pip install. In this way, users can easily install the package with one line of command, without the need to manually manage dependencies and configure the package. We have made the installation instructions clearer as suggested ([6f02afd](https://github.com/caimeng2/seesus/commit/6f02afdd37ddea84227b5faacff62af05000cdde)).

> That being said, I am unable to install and run this package on my machine. I will be happy to do so if one of the authors can help me install the package.

I am more than happy to help. Do you already have Python, pip, and Jupyter (for running the example.ipynb) installed? If yes, typing `pip install seesus` in your terminal should do the job. If not, I would recommend installing Anaconda first. Please go to [Anaconda's website](https://www.anaconda.com/products/distribution) and install it for your specific operating system ([instructions](https://docs.anaconda.com/anaconda/install/)). Then you should be able to install `seesus` by inputting `pip install seesus` in your terminal. Please let me know if you encounter any problems.

**Software paper**

> The idea is novel and I can see this work being useful in many domains. I have one question regarding the example use cases listed in the paper: the paper claims that seesus can be used in "label academic publications" and "large-scale scans of planning documents". However, the example in README.md only shows how seesus evaluates individual sentences which can cause potential misinterpretation and biased results as the context of "academic publications" and "planning documents" will likely be missing when being evaluated sentence by sentence. Example 3 provided [here](https://github.com/caimeng2/seesus/blob/main/example.ipynb), for example is not really a paragraph.

Glad to hear that you find our package novel and potentially useful in many domains. To achieve the best results, we recommend splitting a paragraph or a whole document into individual sentences (i.e., using individual sentences as the basic unit for `seesus` to analyze). This was the reason why we only showed how `seesus` evaluates individual sentences in `README.md` at the beginning. Thank you for pointing out that this might cause misinterpretation. To address this concern, first, we have moved the paragraph example (example 3) in `example.ipynb` to `README.md` ([66862f8](https://github.com/caimeng2/seesus/commit/66862f8a780c548e4252d57e9231e047ab64c922)). [Here](https://github.com/caimeng2/seesus/blob/main/README.md) you can tell this example is a paragraph (i.e., a chunk of text with several sentences) by scrolling to the right. The display of a Jupyter Notebook in GitHub is a bit confusing because the text is truncated. We’ve added a print statement to prevent this confusion ([c77e33e](https://github.com/caimeng2/seesus/commit/c77e33e3bad1db4ed294b9ef7a3108d8a98c4d08)). Second, we have added another example in [`example.ipynb`](https://github.com/caimeng2/seesus/blob/main/example.ipynb) to demonstrate the package’s usage in the context of academic publications ([c77e33e](https://github.com/caimeng2/seesus/commit/c77e33e3bad1db4ed294b9ef7a3108d8a98c4d08)). For both the examples of an academic publication and a planning document, we split the paragraphs into sentences and printed out the results for each sentence. Users can organize the results according to their needs.

> The statement of need is clear but a bit thin. Although I can appreciate JOSS is a more software-focused journal, it would still be great to provide some context on the current status of text mining/classification on UN-SDG and why it is important to for example, "quantify which dimension of sustainability receives the most attention"

Thank you for your suggestion. We have incorporated additional context of text mining on SDGs in our paper as suggested ([c8162fe](https://github.com/caimeng2/seesus/commit/c8162fef21684c396919351e52e5b823facbd599)). Given that JOSS requires papers to be between 250-1000 words ([source](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)), we hope the edits are sufficient to provide the necessary improvement to our statement of need.

> Accuracy of 75.5% is decent but is not particularly high. Considering the evaluation method is from a different package, it would be great if the authors can provide a statement (or even better specific development plans) on how to improve the accuracy and/or usabililty of future text-mining on SDG.

This is a great idea. We’ve added a statement on maintenance in [`README.md`](https://github.com/caimeng2/seesus/blob/main/README.md) to address this ([1eacaaf](https://github.com/caimeng2/seesus/commit/1eacaaf78caf486d4fff5c901757032776a9df51#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5)). Following the best practices of open-source software, we welcome and encourage users to report issues if they find that a matching syntax is not accurate or can be improved.

Thanks again for your time and suggestions!
