### 2024-01-25 #6244 @luyuhao0326
> I am more than happy to help. Do you already have Python, pip, and Jupyter (for running the `example.ipynb`) installed? If yes, typing `pip install seesus` in your terminal should do the job. If not, I would recommend installing Anaconda first. Please go to [Anaconda's website](https://www.anaconda.com/products/distribution) and install it for your specific operating system ([instructions](https://docs.anaconda.com/anaconda/install/)). Then you should be able to install `seesus` by inputting `pip install seesus` in your terminal. Please let me know if you encounter any problems.

Hello, I managed to install the package and while testing [one](https://github.com/caimeng2/seesus?tab=readme-ov-file#analyzing-a-paragraph-longer-document) of the provided examples, I encountered a LookupError. Please see the code [here](http://localhost:8888/notebooks/seesus_review.ipynb).
