### 2024-02-01 #6244 @luyuhao0326
> > Hello, I managed to install the package and while testing [one](https://github.com/caimeng2/seesus?tab=readme-ov-file#analyzing-a-paragraph-longer-document) of the provided examples, I encountered a LookupError. Please see the code [here](http://localhost:8888/notebooks/seesus_review.ipynb).
>
> Hi @luyuhao0326, I'm so glad that you got the installation working 🎉 The link to the LookupError is pointing to your localhost so I can't see the traceback. But I suspect it's a bug with `nltk` (see [this](https://stackoverflow.com/questions/35861482/nltk-lookup-error)). Feel free to try some of the solutions proposed there. An alternative is to use `re` instead of `nltk`. Please see if the following code works.
>
> ```python
> from seesus import SeeSus
> import re
>
> text2 = "By working with communities in the floodplain and facilitating flood-resistant building design, DCP is reducing the city’s risks to sea level rise and coastal flooding. Hurricane Sandy was a stark reminder of these risks. The City, led by the Mayor’s Office of Recovery and Resiliency (ORR), has developed a multifaceted plan for recovering from Sandy and improving the city’s resiliency–the ability of its neighborhoods, buildings and infrastructure to withstand and recover quickly from flooding and climate events. As part of this effort, DCP has initiated a series of projects to identify and implement land use and zoning changes as well as other actions needed to support the short-term recovery and long-term vitality of communities affected by Hurricane Sandy and other areas at risk of coastal flooding."
>
> for sent in re.split(r'(?<!\w\.\w.)(?<![A-Z][a-z]\.)(?<=\.|\?)\s', text2):
>     result = SeeSus(sent)
>     print('"', sent, '"', sep = "")
>     print("Is the sentence related to achieving sustainability?", result.sus)
>     print("Which SDGs?", result.sdg)
>     print("Which SDG targets specifically?", result.target)
>     print("which dimensions of sustainability?", result.see)
>     print("----------------")
> ```
>
> Thank you for letting me know about this issue. I'll update the examples.

Indeed this is the bug. It is now working with `re`
