### 2024-02-19 #6244 @varsha2509
Hello. Thank you for giving me the opportunity to review this work. This authors have done a great job documenting the software, installation instructions and the functionality. Below are my comments based on the review checklist as well as a additional notes to help improve readability and adoption of this work:

1. **Functionality** - while the functional claims of the software have been verified, could the authors provide more details on how the indirect search keys in `SDG_keys.py` specifically from [this](https://github.com/caimeng2/seesus/blob/eeea37e2ef18fb378aff5246b6c7fa15ed0f24bd/src/seesus/SDG_keys.py#L189) line onwards were determined? Can the authors confirm that all targets were included in the keywords?
2. **Automated tests** - while the existing tests cover the explained functionality, the authors should consider including more examples in tests, especially relevant sentences that have a negative connotation to clarify the performance of this tool.
3. **Statement of need** -
- The existing statement of need isn't particularly strong. It's not very clear to me what the benefits of `seesus` are over existing tools, other than the functionality which classifies the expression as environmental, social or economic
sustainability. Making the statement of need stronger will help improve adoption of this tool.
- Could the authors provide an example of what they mean by "also the attainment of SDGs" as specified in the statement of need?
4. **State of the field** -
- OSDG (https://arxiv.org/abs/2211.11252, https://github.com/osdg-ai/osdg-tool) is another open source tool for text based classification of SDG goals and these use NLP/ML based methods. This may be worth highlighting as one of the other classifiers in the statement of need. Along with this, could the authors also include why users would consider `seesus` over existing open source tools?

Other notes:
- Running through the code and script as examples, the current tool is not able to capture negative expressions, as Regex lacks semantic understanding of text. For instance using this sentence as an input "One should not resolve climate change for environmental sustainability." this is being classified as relating to achieving SDG13 and SDG15 but the output should be 'None' or 'Does not match SDG goals'.
This seems to be a limitation of this tool and it would be worth highlighting this in a separate section and including some ideas on how the authors plan to address these limitations in future releases of this tool. This will help users be fully aware of the benefits and limitations of this software.
- Related to above, could the authors talk briefly about limitations of regex for pattern matching over existing semantic text search/language models?
- The authors mention that `seesus` achieves an accuracy rate of 75.5%, as determined by alignment with manual coding. Can the authors comment on how they plan to improve the performance of this tool in future releases as 75% accuracy currently seems low for usability.
