+++ tests/test_seesus.py
@@ -5,12 +5,18 @@ from seesus import seesus
 
 def test_id_sus():
     """Test identifying the UN Sustainable Development Goals (SDGs) and their associated targets in text."""
-    text = "our goal is end poverty"
-    sus, sdgs, targets, matches = seesus.id_sus(text)
-    assert sus == True
-    assert sdgs == ['SDG1']
-    assert targets == ['SDG1_2']
-    assert matches == ['indirect']
+    text1 = "our goal is to end poverty"
+    text2 = "we care about SDG 2"
+    sus1, sdgs1, targets1, matches1 = seesus.id_sus(text1)
+    sus2, sdgs2, targets2, matches2 = seesus.id_sus(text2)
+    assert sus1 == True
+    assert sdgs1 == ['SDG1']
+    assert targets1 == ['SDG1_2']
+    assert matches1 == ['indirect']
+    assert sus2 == True
+    assert sdgs2 == ['SDG2']
+    assert targets2 == ['SDG2_general']
+    assert matches2 == ['direct']
 
 def test_cat_sus():
     """Test categorizing SDG targets into social, environmental, and economic sustainability."""
