+++ paper/paper.bib
@@ -70,6 +70,19 @@
  pages={2819–2830}
 }
 
+@article{Pukelis_Bautista-Puig_Statulevičiūtė_Stančiauskas_Dikmener_Akylbekova_2022,
+ title={OSDG 2.0: a multilingual tool for classifying text data by UN Sustainable Development Goals (SDGs)},
+ url={http://arxiv.org/abs/2211.11252},
+ DOI={10.48550/arXiv.2211.11252},
+ abstractNote={Despite concrete indicators and targets, monitoring the progress of the UN Sustainable Development Goals (SDGs) remains a challenge, given the many different actors, initiatives, and institutions involved. OSDG, an open-source classification tool aims to help navigate the SDG related ambiguities through a simple and easy to use application. The tool allows to map and connect activities to the SDGs by identifying SDG -relevant content in any text. This paper presents OSDG 2.0, a new iteration of the partnership’s work, which marks a significant improvement in the tool’s methodology, as well as support for content in 15 languages.},
+ note={arXiv:2211.11252 [cs]},
+ number={arXiv:2211.11252},
+ publisher={arXiv},
+ author={Pukelis, Lukas and Bautista-Puig, Nuria and Statulevičiūtė, Gustė and Stančiauskas, Vilius and Dikmener, Gokhan and Akylbekova, Dina},
+ year={2022},
+ month=nov 
+}
+
 @misc{Rawat_2022,
  type={Jupyter Notebook},
  title={SDG-Classifier},
