### 2024-02-21 #6244 @caimeng2
Hi @varsha2509,

Thank you for taking the time to review our software and for your valuable feedback. Please find our point-by-point responses below.

>Functionality - while the functional claims of the software have been verified, could the authors provide more details on how the indirect search keys in SDG_keys.py specifically from [this](https://github.com/caimeng2/seesus/blob/eeea37e2ef18fb378aff5246b6c7fa15ed0f24bd/src/seesus/SDG_keys.py#L189) line onwards were determined? Can the authors confirm that all targets were included in the keywords?

Yes, we can confirm that all targets are included in the keywords. We created the search keys at the levels of both the 17 SDGs and the 169 SDG targets. The indirect keys were first based on Thesaurus, and we (four researchers specialized in SDGs) manually assessed and improved the accuracy of the matching syntax by using thousands of randomly-selected statements from corporate reports. We conducted three rounds of fine-tuning and finalized these keys.

>Automated tests - while the existing tests cover the explained functionality, the authors should consider including more examples in tests, especially relevant sentences that have a negative connotation to clarify the performance of this tool.

Thank you for pointing this out. Indeed, matching with negative connotation is `seesus`’s limitation. `seesus` can identify the terms related to SDGs but cannot distinguish between achieving SDGs and failing to do so. This limitation lies in regular expression’s limited logic capability and lack of context awareness. We have added another test of direct matching ([4968c35](https://github.com/caimeng2/seesus/commit/4968c3553c3069e47fd56ac87bf47a4070270e0e)) and edited the paper, deleting expressions regarding “attainment of SDGs,” to make it clear that `seesus` is designed to classify based on relevance.

>Statement of need -
The existing statement of need isn't particularly strong. It's not very clear to me what the benefits of seesus are over existing tools, other than the functionality which classifies the expression as environmental, social or economic
sustainability. Making the statement of need stronger will help improve adoption of this tool.

The biggest benefit of `seesus` is its finer scale: it captures not only the SDGs but also the 169 SDG targets. To the best of our knowledge, no other Python tool does this. In addition, compared to tools based on machine learning, `seesus` allows users to examine and modify the matching syntax, so users can always understand and have control over the results. We’ve edited the statement of need to make it stronger as suggested ([3f35864](https://github.com/caimeng2/seesus/commit/3f35864e17a665d89147ee69ba26600f2529f398)).

>Could the authors provide an example of what they mean by "also the attainment of SDGs" as specified in the statement of need?

What we meant was `seesus` specifically looks for terms that are related to achieving the SDGs, and not just SDG-related topics themselves. For example, it is not to find words solely related to emissions (e.g., "emissions", "carbon"), but it looks for terms such as "lowering emissions" and "reducing carbon." However, we realized that this sentence is rather confusing as `seesus` cannot identify negative expressions, so we have deleted it to avoid further confusion.

>State of the field -
OSDG (https://arxiv.org/abs/2211.11252, https://github.com/osdg-ai/osdg-tool) is another open source tool for text based classification of SDG goals and these use NLP/ML based methods. This may be worth highlighting as one of the other classifiers in the statement of need. Along with this, could the authors also include why users would consider seesus over existing open source tools?

Thank you for this reference. We have added it to the existing classifiers. We tested `OSDG` and noticed that it is not able to capture negative expressions either, and the results are only the 17 SDGs, not the targets. We have revised our paper to highlight `seesus`’s benefit.

>Other notes: Running through the code and script as examples, the current tool is not able to capture negative expressions, as Regex lacks semantic understanding of text. For instance using this sentence as an input "One should not resolve climate change for environmental sustainability." this is being classified as relating to achieving SDG13 and SDG15 but the output should be 'None' or 'Does not match SDG goals'.
Screenshot 2024-02-19 at 3 23 50 PM This seems to be a limitation of this tool and it would be worth highlighting this in a separate section and including some ideas on how the authors plan to address these limitations in future releases of this tool. This will help users be fully aware of the benefits and limitations of this software.
Related to above, could the authors talk briefly about limitations of regex for pattern matching over existing semantic text search/language models?

Thank you for your suggestion. This is a very good point. Compared to language models, regex lacks the ability to understand the semantic meaning or context of text, as it operates based on character patterns. As suggested, we have added a paragraph at the end of the paper to make the limitation of `seesus` more clear.

>The authors mention that seesus achieves an accuracy rate of 75.5%, as determined by alignment with manual coding. Can the authors comment on how they plan to improve the performance of this tool in future releases as 75% accuracy currently seems low for usability.

Yes, we have included it in the revision of the paper. We devoted hundreds of hours to fine-tune the matching syntax. 75.5% seems low but it is quite reasonable for traditional qualitative analysis. The human intercoder agreement on the same text was only at 83%.

Thanks again for all your comments and suggestions! We feel that our paper is much clearer and stronger than the previous version. Thank you!!! Please let us know if there's anything else.
