### 2024-03-03 #6244 @varsha2509
Hello @caimeng2  - thank you for addressing my comments. Please see my responses below:

> Yes, we can confirm that all targets are included in the keywords. We created the search keys at the levels of both the 17 SDGs and the 169 SDG targets. The indirect keys were first based on Thesaurus, and we (four researchers specialized in SDGs) manually assessed and improved the accuracy of the matching syntax by using thousands of randomly-selected statements from corporate reports. We conducted three rounds of fine-tuning and finalized these keys.
Thanks for confirming this. Depending on word limit, I'd recommend include a line or two about this in the paper, or in the Github Readme, under a methodology section.

Besides the suggestion above, the authors have fully addressed all of my comments and made revisions where necessary. I recommend this submission to be accepted by JOSS.
