### 2024-03-05 522f731
shorten methodology
