### 2024-03-05 716d675
shorten methodology
