### 2024-03-05 #6244 @caimeng2
Hi @varsha2509,

>Thanks for confirming this. Depending on word limit, I'd recommend including a line or two about this in the paper, or in the Github Readme, under a methodology section.

This is a great suggestion! We revised the paper accordingly ([018e7e1](https://github.com/caimeng2/seesus/commit/018e7e1eea61af2aaeb3e249113572b4513b9772)) and added a methodology section to [README](https://github.com/caimeng2/seesus/blob/main/README.md).

We are glad to hear that you found our revisions satisfactory, and appreciate your recommendation for acceptance. Thank you again for your thorough review of our submission!
