### 2024-03-22 #6244 @caimeng2
Hi @oliviaguest, I finished the author tasks listed above. Please let me know if there's anything else :nerd_face:

> * Double check authors and affiliations (including ORCIDs)

Checked

> * Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

[v1.0](https://github.com/caimeng2/seesus/releases/tag/v1.0)

> * Archive the release on Zenodo/figshare/etc and post the DOI here.

[![DOI](https://zenodo.org/badge/620332547.svg)](https://zenodo.org/doi/10.5281/zenodo.10854083)

> * Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

Checked

> * Make sure that the license listed for the archive is the same as the software license.

Checked
