### 2024-03-26 #6244 @caimeng2
> @caimeng2 why is it `Version: v1.2.0` above?

Ah my bad. That was the version number for PyPI, which I totally forgot. Should have made them consistent
