### 2024-03-26 #6244 @caimeng2
Hi @oliviaguest, I made a new release and redid the tasks above. Sorry about the inconvenience.

> Double check authors and affiliations (including ORCIDs)

Checked

> Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

[v1.2.1](https://github.com/caimeng2/seesus/releases/tag/v1.2.1)

> Archive the release on Zenodo/figshare/etc and post the DOI here.

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10878371.svg)](https://doi.org/10.5281/zenodo.10878371)

> Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

Checked

> Make sure that the license listed for the archive is the same as the software license.

Checked
