### 2024-04-07 #6244 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
